ImageSmoothening/requirements.txt
torch
torchvision
pillow
numpy
pandas
